+++ NEWS.md
@@ -1,4 +1,9 @@
 
+### fitbitScraper 0.1.4
+* added get_activity_data() function
+* added "getRestingHeartRateData" to get_daily_data() function
+* added rememberMe parameter to login function
+* merged pull request #2, a change to the login function
 
 ### fitbitScraper 0.1.3
 * Changed the API calls to match changes on fitbit end of things.